fix: Dynamic require of "fs" is not supported. fixes #13
This commit moves dotenv and dotenv-expand into `dependencies` from `devDependencies`, which lets tsup externalize them.
Therefore, in the bundled code, they will be imported using native `import`s instead of `__require`s generated by esbuild, which in turn tries to call native `require` on NodeJS, which is not supported in NodeJS when the file is treated as an ESM. The detailed cause is clear in https://github.com/evanw/esbuild/issues/1921.